matlab bindings working on vs 2010 export and importhdf5RF are not working due to some linker errors (error LNK2005: _malloc_crt already defined in MSVCRT.lib)

+++ include/vigra/random_forest/splices.hxx
@@ -76,7 +76,7 @@ class Splice<int>
     : begin_(begin), 
       interval_(interval),
       end_(end),
-      size_(std::floor((double(end) -double(begin))/interval))
+      size_(int(std::floor((double(end) -double(begin))/interval)))
     {}
     
     int operator[](int index)
